Updating install script

+++ script/uninstall.sh
@@ -6,11 +6,19 @@
 
 # Temporary "uninstall" for testing
 cd $HOME
-#mv .oh-my-zsh .oh-my-zsh.bak
-#mv dotfiles dotfiles.bak
 
-rm -rf dotfiles
-rm -rf .oh-my-zsh
+backup=true
+
+if [ $backup = true ]; then
+
+  mv .oh-my-zsh .oh-my-zsh.bak
+  mv dotfiles dotfiles.bak
+  mv .rvm .rvm.bak
+else
+  rm -rf dotfiles
+  rm -rf .oh-my-zsh
+  rm -rf .rvm
+fi
+
 rm .zshrc
-
 chsh -s /bin/bash
